# apps/backend/src/devices/dto/register-device.dto.ts
import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
import { IsOptional, IsString, MinLength } from 'class-validator';

export class RegisterDeviceDto {
  @ApiProperty({ description: 'Reusable per-location setup code generated by the kiosk owner/admin' })
  @IsString()
  @MinLength(1)
  setupCode: string;

  @ApiPropertyOptional({ example: 'Computer 4' })
  @IsOptional()
  @IsString()
  label?: string;

  @ApiPropertyOptional({ description: 'Falls back to this if no label is given' })
  @IsOptional()
  @IsString()
  hostname?: string;
}