# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login succeeded"
          },
          "401": {
            "description": "Invalid credentials, or account disabled"
          }
        },
        "summary": "Log in with email + password, get an access/refresh token pair",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A fresh token pair; the old refresh token is now invalid"
          },
          "401": {
            "description": "Invalid, expired, or already-rotated refresh token"
          }
        },
        "summary": "Exchange a refresh token for a new access/refresh token pair",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logged out"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Invalidate the current refresh token",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/kiosks/{id}": {
      "get": {
        "operationId": "KiosksController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The kiosk"
          },
          "403": {
            "description": "Not this kiosk's owner and not an admin"
          },
          "404": {
            "description": "Kiosk not found"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Get a kiosk by id (admin, or the owning kiosk-owner)",
        "tags": [
          "Kiosks"
        ]
      },
      "patch": {
        "operationId": "KiosksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateKioskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Kiosk updated"
          },
          "404": {
            "description": "Kiosk not found"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Update a kiosk, including revenueSharePct and status (admin only)",
        "tags": [
          "Kiosks"
        ]
      }
    },
...
    "/kiosks/{kioskId}/users": {
      "post": {
        "description": "Admin can create KIOSK_OWNER or LOCATION_MANAGER accounts. A kiosk-owner may only create LOCATION_MANAGER accounts under their own kiosk — never a peer owner.",
        "operationId": "KioskUsersController_create",
        "parameters": [
          {
            "name": "kioskId",
            "required": true,
            "in": "path",
            "description": "The kiosk these user accounts belong to",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKioskUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          },
          "403": {
            "description": "Cross-tenant, or a kiosk-owner attempting to assign a disallowed role"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Create a user under this kiosk",
        "tags": [
          "Kiosk Users"
        ]
      },
      "get": {
        "operationId": "KioskUsersController_findAll",
        "parameters": [
          {
            "name": "kioskId",
            "required": true,
            "in": "path",
            "description": "The kiosk these user accounts belong to",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users under this kiosk (password hashes never included)"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "List all users belonging to this kiosk",
        "tags": [
          "Kiosk Users"
        ]
      }
    },
    "/kiosks/{kioskId}/users/{userId}": {
      "patch": {
        "operationId": "KioskUsersController_update",
        "parameters": [
          {
            "name": "kioskId",
            "required": true,
            "in": "path",
            "description": "The kiosk these user accounts belong to",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateKioskUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated"
          },
          "403": {
            "description": "Cross-tenant, or a kiosk-owner acting on a non-location-manager account"
          },
          "404": {
            "description": "User does not belong to this kiosk"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Update a kiosk user — role, disabled flag, or managedLocationIds",
        "tags": [
          "Kiosk Users"
        ]
      }
    },
...
    "/devices/register": {
      "post": {
        "description": "No human auth — gated by a valid, active LocationSetupCode whose parent kiosk is ACTIVE. Returns a device token immediately, no approval step. Rate-limited (5/min) to blunt setup-code guessing.",
        "operationId": "DevicesController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Device registered; token returned once, store it — it is not retrievable again"
          },
          "400": {
            "description": "Setup code invalid/inactive, or parent kiosk not ACTIVE"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "summary": "Machine-initiated device registration",
        "tags": [
          "Devices"
        ]
      }
    },
    "/devices": {
      "get": {
        "description": "Admin sees all devices; kiosk-owner sees their own kiosk's devices; location-manager sees only devices at their assigned locations.",
        "operationId": "DevicesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Devices visible to the caller"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "List devices",
        "tags": [
          "Devices"
        ]
      }
    },
    "/devices/{id}": {
      "get": {
        "operationId": "DevicesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The device"
          },
          "403": {
            "description": "Not in the caller's tenant scope"
          },
          "404": {
            "description": "Device not found"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Get a device by id",
        "tags": [
          "Devices"
        ]
      },
      "patch": {
        "operationId": "DevicesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDeviceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Device updated — active:false blocks its token on next use"
          },
          "403": {
            "description": "Not in the caller's tenant scope"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Update a device, including the active kill-switch",
        "tags": [
          "Devices"
        ]
      },
      "delete": {
        "operationId": "DevicesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Device deleted"
          },
          "403": {
            "description": "Not allowed for this role or tenant"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Delete a device and revoke its tokens (admin or kiosk-owner only)",
        "tags": [
          "Devices"
        ]
      }
    }
  },
  "info": {
    "title": "Saverlly API",
    "description": "Multi-tenant SaaS platform for internet kiosk businesses — admin console, kiosk portal, and machine (Chrome extension / desktop agent) endpoints.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "device-token": {
        "scheme": "bearer",
        "bearerFormat": "device-token",
        "type": "http",
        "description": "Opaque device token issued at /devices/register — used by machine clients only."
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "owner@kiosk.com"
          },
          "password": {
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "UpdateKioskDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "revenueSharePct": {
            "type": "number",
            "description": "Percentage of commission the kiosk keeps (0-100)"
          },
          "contactEmail": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          }
        }
      },
...
      "CreateKioskUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "writeOnly": true,
            "minLength": 8
          },
          "role": {
            "type": "string",
            "enum": [
              "KIOSK_OWNER",
              "LOCATION_MANAGER"
            ]
          },
          "managedLocationIds": {
            "description": "Location ids this user manages — only meaningful for LOCATION_MANAGER",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "email",
          "password",
          "role"
        ]
      },
      "UpdateKioskUserDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "KIOSK_OWNER",
              "LOCATION_MANAGER"
            ]
          },
          "disabled": {
            "type": "boolean",
            "description": "Disable/re-enable this user account"
          },
          "managedLocationIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "RegisterDeviceDto": {
        "type": "object",
        "properties": {
          "setupCode": {
            "type": "string",
            "description": "Reusable per-location setup code generated by the kiosk owner/admin"
          },
          "label": {
            "type": "string",
            "example": "Computer 4"
          },
          "hostname": {
            "type": "string",
            "description": "Falls back to this if no label is given"
          }
        },
        "required": [
          "setupCode"
        ]
      },
      "UpdateDeviceDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          },
          "active": {
            "type": "boolean",
            "description": "Kill-switch — false immediately blocks this device's token"
          }
        }
      }
    }
  }
}